requirements.txt
fastapi
uvicorn
httpx
anthropic
python-dotenv
pydantic
pydantic-settings
